+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add `Launcher.addPlayerName(playerName: string)`
 
 ## [0.9.1] - 2016-11-19
 ### Fixed